backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
langgraph==0.0.62
langchain==0.1.0
langchain-openai==0.0.2
openai==1.3.9
google-cloud-vision==3.4.5
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.4
pandas==2.1.4
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1